Re-add missing xcodeproj file
