+++ 03-TimeConversionApp/TimeConverter/TimeConverter/ContentView.swift
@@ -1,0 +1,26 @@
+//
+//  ContentView.swift
+//  TimeConverter
+//
+//  Created by Phil on 09/02/2023.
+//
+
+import SwiftUI
+
+struct ContentView: View {
+    var body: some View {
+        VStack {
+            Image(systemName: "globe")
+                .imageScale(.large)
+                .foregroundColor(.accentColor)
+            Text("Hello, world!")
+        }
+        .padding()
+    }
+}
+
+struct ContentView_Previews: PreviewProvider {
+    static var previews: some View {
+        ContentView()
+    }
+}
